added gui utils, started remote api, readme change
